+++ .github/workflows/publish.yml
@@ -10,6 +10,7 @@ jobs:
     strategy:
       matrix:
         package:
+          - abi
           - polkadart
           - polkadart_cli
           - polkadart_keyring
