+++ packages/ink_abi/example/v3_metadata.json
