+++ packages/ink_abi/lib/ink_abi_base.dart
@@ -1,4 +1,4 @@
-library abi;
+library ink_abi;
 
 import 'dart:typed_data';
 
@@ -7,11 +7,11 @@ import 'package:polkadart_scale_codec/polkadart_scale_codec.dart';
 import 'package:substrate_metadata/parsers/parsers.dart';
 
 // core
-part 'core/abi.dart';
-part 'core/abi_description.dart';
+part 'core/ink_abi.dart';
+part 'core/ink_abi_description.dart';
 
 // models
-part 'models/abi_event.dart';
+part 'models/ink_abi_event.dart';
 
 // schemas
 part 'schemas/v3/ink_v3_schema.dart';
