+++ packages/ink_abi/lib/utils/definitions.dart
@@ -1,3 +1,3 @@
-part of abi;
+part of ink_abi;
 
 typedef SelectorsMap = Map<String, int>;
