+++ packages/ink_abi/test/erc_20_v3_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:abi/abi_base.dart';
+import 'package:ink_abi/ink_abi_base.dart';
 import 'package:polkadart_scale_codec/polkadart_scale_codec.dart';
 import 'package:test/test.dart';
 
@@ -9,11 +9,11 @@ void main() {
   final String jsonFilePath = './test/test_resources/v3_metadata.json';
   final Map<String, dynamic> metadataV3 =
       jsonDecode(File(jsonFilePath).readAsStringSync());
-  final Abi abi = Abi(metadataV3);
+  final InkAbi inkAbi = InkAbi(metadataV3);
 
   test('decode constructor', () {
     final data = '0x9bae9d5e0000a0dec5adc9353600000000000000';
-    final decoded = abi.decodeConstructor(data);
+    final decoded = inkAbi.decodeConstructor(data);
     expect(decoded.value['initial_supply'].toString(),
         BigInt.tryParse('1000000000000000000000').toString());
   });
@@ -21,7 +21,7 @@ void main() {
   test('decode event', () {
     final String data =
         '0x0001da002226d93b2c422b95b780a2493e738716050ccad6ddbd7d58f1943bc6373d015207202c27b646ceeb294ce516d4334edafbd771f869215cb070ba51dd7e2c720000c84e676dc11b0000000000000000';
-    final decoded = abi.decodeEventFromHex(data);
+    final decoded = inkAbi.decodeEventFromHex(data);
     expect('0x${encodeHex(decoded['from'].value.cast<int>())}',
         '0xda002226d93b2c422b95b780a2493e738716050ccad6ddbd7d58f1943bc6373d');
     expect('0x${encodeHex(decoded['to'].value.cast<int>())}',
@@ -33,7 +33,7 @@ void main() {
   test('decode message', () {
     final String data =
         '0x84a15da15207202c27b646ceeb294ce516d4334edafbd771f869215cb070ba51dd7e2c720000f444829163450000000000000000';
-    final decoded = abi.decodeMessage(data);
+    final decoded = inkAbi.decodeMessage(data);
     expect('0x${encodeHex(decoded.value['to'].cast<int>())}',
         '0x5207202c27b646ceeb294ce516d4334edafbd771f869215cb070ba51dd7e2c72');
     expect(decoded.value['value'].toString(),
