+++ packages/ink_abi/test/erc_20_v5_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:abi/abi_base.dart';
+import 'package:ink_abi/ink_abi_base.dart';
 import 'package:polkadart_scale_codec/polkadart_scale_codec.dart';
 import 'package:test/test.dart';
 
@@ -9,11 +9,11 @@ void main() {
   final String jsonFilePath = './test/test_resources/v5_metadata.json';
   final Map<String, dynamic> metadataV5 =
       jsonDecode(File(jsonFilePath).readAsStringSync());
-  final Abi abi = Abi(metadataV5);
+  final InkAbi inkAbi = InkAbi(metadataV5);
 
   test('decode constructor', () {
     final String data = '0x9bae9d5e00000c6d51c8f7aa0600000000000000';
-    final decoded = abi.decodeConstructor(data);
+    final decoded = inkAbi.decodeConstructor(data);
     expect(decoded.value['total_supply'].toString(),
         BigInt.tryParse('123000000000000000000').toString());
   });
@@ -26,7 +26,7 @@ void main() {
       '0xda002226d93b2c422b95b780a2493e738716050ccad6ddbd7d58f1943bc6373d',
       '0xa69162c917081d15673558e13607b1b2261f2ae7b21ba911c3cd676767251266'
     ];
-    final decoded = abi.decodeEventFromHex(data, topics);
+    final decoded = inkAbi.decodeEventFromHex(data, topics);
     expect('0x${encodeHex(decoded['owner'].cast<int>())}',
         '0xda002226d93b2c422b95b780a2493e738716050ccad6ddbd7d58f1943bc6373d');
     expect('0x${encodeHex(decoded['spender'].cast<int>())}',
@@ -42,7 +42,7 @@ void main() {
       '0xda002226d93b2c422b95b780a2493e738716050ccad6ddbd7d58f1943bc6373d',
       '0xa69162c917081d15673558e13607b1b2261f2ae7b21ba911c3cd676767251266'
     ];
-    final decoded = abi.decodeEventFromHex(data, topics);
+    final decoded = inkAbi.decodeEventFromHex(data, topics);
     expect(('0x${encodeHex(decoded['from'].value.cast<int>())}'),
         '0xda002226d93b2c422b95b780a2493e738716050ccad6ddbd7d58f1943bc6373d');
     expect(('0x${encodeHex(decoded['to'].value.cast<int>())}'),
@@ -54,7 +54,7 @@ void main() {
   test('decode message', () {
     final String data =
         '0x84a15da1a69162c917081d15673558e13607b1b2261f2ae7b21ba911c3cd67676725126600008a5d784563010000000000000000';
-    final decoded = abi.decodeMessage(data);
+    final decoded = inkAbi.decodeMessage(data);
     expect(('0x${encodeHex(decoded.value['to'].cast<int>())}'),
         '0xa69162c917081d15673558e13607b1b2261f2ae7b21ba911c3cd676767251266');
     expect(decoded.value['value'].toString(),
