+++ packages/ink_abi/test/test_resources/v3_metadata.json
