+++ packages/ink_abi/test/test_resources/v5_metadata.json
